reorganized project structure to make space for expression-parser
