--- src/tokens.ts
@@ -1,7 +1,0 @@
-export const Tag = {
-  Assignment: ["="],
-  Close: [">", "/>"],
-  Open: ["<"],
-  SelfClose: ["/>"],
-  String: [`'`, `"`],
-};
