+++ src/transpiler.ts
@@ -1,4 +1,4 @@
-import * as KNode from "./nodes";
+import * as KNode from "@kasper/types/nodes";
 
 export class Transpiler implements KNode.NodeVisitor<Node> {
   public errors: string[] = [];
